[Feat] Vinyls Respository Interface 작성 및 관련 remote Service, Response code 작성

+++ data/src/main/java/com/vinyla_android/data/remote/service/response/VinylResponse.kt
@@ -14,7 +14,7 @@ data class VinylResponse(
     val artist: String,
 ) {
     fun toVinyl(): Vinyl = Vinyl(
-        id = vinylIdx.toLong(),
+        id = vinylIdx,
         name = title,
         artiestName = artist,
         genre = "",
